Make optimized, unsafe deallocs optional
snmalloc allocators offer two optimized "dealloc" functions which take
(statically or dynamically) the size of the pointer to be freed.
However, these are unsafe in that they can cause type confusion within
snmalloc.  Add two configuration switches: one which causes these unsafe
interfaces to take time to assert that they are being used correctly and
one which simply shunts the unsafe call to the slower, safer form which
which rederives the size from snmalloc-internal metadata.