CHERI: don't bound queue-priming stubs
These objects never leave the allocator and get fed directly to the
internal dealloc()-ation functions, so it's simplest to leave them
unrestricted rather than having to rederive the unrestricted form later.
For non-CHERI architectures, there is no change.

When we come to do temporally-safe revocation, our current design has
all objects on the queue being the internal capabilities because
quarantine happens on the initial dealloc()-ating side, not within the
recipient.