CHERI: de-weaponize external_address
If we're not careful, external_address provides a path for a user
pointer to survive revocation, since the natural thing to do is promote
the pointer the user gives us and then reason entirely from there and
return a (bounded) result.  Instead, ensure that our use of rederived
capabilities is strictly internal and that our result is always derived
from the pointer given by the user, thereby closing any TOCTTOU race
with a concurrent revocation pass.