realloc vs. CHERI
Rederive caps even when the data aren't moving: we bound to the original
allocation size, not the sizeclass's size (to impose more strict
requirements on our caller).

Ensure that not only the requested size is of the same sizeclass, but
also the representable padded version of the caller's requested size.

This is a variant of a fix proposed by Lucian Paul-Trifu, who found the
issue in the first place.